Create DIGF2B03 Physical Computing Lab 3 Question 3
This code reads values from an LDR to determine the lighting of the atmosphere. If the lights are off, a speaker will start playing "Twinkle Twinle Little Star", and an LED will light up to a certain brightness depending on how dark the room is. If there is light, the LED turns off, and the sound stops playing. The song continues from where it left off when the room is dark again.